requirements.txt
ghapi==1.0.4
mkdocs-material==9.2.0b3
mkdocs-macros-plugin==1.0.2
mkdocs-include-markdown-plugin==5.1.0
requests==2.31.0
